Fixed runtime analysis error:  left shift of negative value
uint8_t ^ uint8_t is promoted to signed int. Then when the complement is taken we arrive at a negative result, which is then left shifted, which is an undefined operation

git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38572 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/iecbus/iecbus.c
@@ -265,7 +265,7 @@ static void iecbus_cpu_write_conf1(uint8_t data, CLOCK clock)
         default:
             iecbus.drv_bus[8] = (((iecbus.drv_data[8] << 3) & 0x40)
                                  | ((iecbus.drv_data[8] << 6)
-                                    & ((~iecbus.drv_data[8] ^ iecbus.cpu_bus) << 3)
+                                    & ((uint32_t)(~iecbus.drv_data[8] ^ iecbus.cpu_bus) << 3)
                                     & 0x80));
     }
     iec_update_ports();
